feat: move Breakout71 game-specific code to games/breakout71/ plugin directory
Move Breakout71Env, Breakout71Loader, JS modal constants, and YOLO class
names into games/breakout71/ to separate game-specific code from the
platform. Backward-compat re-exports in src/env/ and src/game_loader/
use lazy __getattr__ to avoid circular imports. Deduplicate JS snippets
from 3 files into single-source-of-truth modal_handler.py. Config files
remain in configs/ (consumed by platform-level loaders).
