+++ games/__init__.py
@@ -1,0 +1,10 @@
+"""Game plugins for the RSN Game QA platform.
+
+Each subdirectory under ``games/`` is a self-contained game plugin
+providing the game-specific glue code: environment subclass, loader,
+modal handler JS snippets, perception class list, and config files.
+
+Current plugins:
+
+- :mod:`games.breakout71` -- Breakout 71 browser game
+"""
