+++ scripts/capture_dataset.py
@@ -55,149 +55,13 @@ STATE_GAME_OVER = "game_over"
 STATE_MENU = "menu"
 STATE_UNKNOWN = "unknown"
 
-# JavaScript snippet that queries the game DOM to determine current state.
-# Returns a JSON-serializable dict with state info.
-_DETECT_STATE_JS = """
-return (function() {
-    var result = {state: "unknown", details: {}};
-
-    // Check if any modal/alert is open
-    var hasAlert = document.body.classList.contains('has-alert-open');
-    var popup = document.getElementById('popup');
-    var closeBtn = document.getElementById('close-modale');
-
-    if (!hasAlert) {
-        // No modal open — treat as gameplay.  We cannot distinguish
-        // paused vs. running from the DOM alone, but the bot's periodic
-        // paddle actions and canvas clicks keep the game unpaused.
-        result.state = "gameplay";
-        return result;
-    }
-
-    // Modal is open — determine which type
-    if (!popup) {
-        result.state = "unknown";
-        return result;
-    }
-
-    var popupText = popup.innerText || "";
-    var buttons = popup.querySelectorAll('button');
-    var buttonTexts = [];
-    for (var i = 0; i < buttons.length; i++) {
-        buttonTexts.push(buttons[i].innerText.trim());
-    }
-    result.details.buttonTexts = buttonTexts;
-    result.details.popupTextSnippet = popupText.substring(0, 200);
-
-    // Check if close button is visible (game-over and menu modals have it;
-    // perk picker does NOT)
-    var closeBtnVisible = false;
-    if (closeBtn) {
-        var style = window.getComputedStyle(closeBtn);
-        closeBtnVisible = (style.display !== 'none' && style.visibility !== 'hidden');
-    }
-    result.details.closeBtnVisible = closeBtnVisible;
-
-    // Perk picker: required modal (no close button), has upgrade buttons
-    if (!closeBtnVisible && buttons.length >= 2) {
-        result.state = "perk_picker";
-        result.details.numPerks = buttons.length;
-        return result;
-    }
-
-    // Game over: has close button, popup text usually contains score info
-    // and a "New Run" / restart button
-    if (closeBtnVisible) {
-        // Check for restart-like button text
-        var hasRestart = false;
-        for (var j = 0; j < buttonTexts.length; j++) {
-            var t = buttonTexts[j].toLowerCase();
-            if (t.indexOf("new") >= 0 || t.indexOf("restart") >= 0 ||
-                t.indexOf("run") >= 0 || t.indexOf("yes") >= 0 ||
-                t.indexOf("again") >= 0) {
-                hasRestart = true;
-                break;
-            }
-        }
-
-        if (hasRestart || popupText.toLowerCase().indexOf("game over") >= 0 ||
-            popupText.toLowerCase().indexOf("score") >= 0) {
-            result.state = "game_over";
-        } else {
-            result.state = "menu";
-        }
-        return result;
-    }
-
-    // Fallback: unknown modal
-    result.state = "unknown";
-    return result;
-})();
-"""
-
-# JavaScript snippet to click a random perk button in the upgrade picker.
-# Returns the index of the button clicked, or -1 if none found.
-_CLICK_PERK_JS = """
-return (function() {
-    var popup = document.getElementById('popup');
-    if (!popup) return {clicked: -1, text: ""};
-
-    var buttons = popup.querySelectorAll('button');
-    if (buttons.length === 0) return {clicked: -1, text: ""};
-
-    // Pick a random button
-    var idx = Math.floor(Math.random() * buttons.length);
-    var text = buttons[idx].innerText.trim();
-    buttons[idx].click();
-    return {clicked: idx, text: text};
-})();
-"""
-
-# JavaScript snippet to dismiss game-over modal.
-# Strategy: try restart-like buttons first, fall back to the close button.
-_DISMISS_GAME_OVER_JS = """
-return (function() {
-    var popup = document.getElementById('popup');
-    var closeBtn = document.getElementById('close-modale');
-
-    // Try clicking any restart-like button first
-    if (popup) {
-        var buttons = popup.querySelectorAll('button');
-        for (var i = 0; i < buttons.length; i++) {
-            var t = buttons[i].innerText.trim().toLowerCase();
-            if (t.indexOf("new") >= 0 || t.indexOf("restart") >= 0 ||
-                t.indexOf("run") >= 0 || t.indexOf("yes") >= 0 ||
-                t.indexOf("again") >= 0) {
-                buttons[i].click();
-                return {action: "restart_button", text: buttons[i].innerText.trim()};
-            }
-        }
-    }
-
-    // Fall back to close button
-    if (closeBtn) {
-        closeBtn.click();
-        return {action: "close_button", text: ""};
-    }
-
-    return {action: "none", text: ""};
-})();
-"""
-
-# JavaScript snippet to dismiss a generic menu modal.
-_DISMISS_MENU_JS = """
-return (function() {
-    var closeBtn = document.getElementById('close-modale');
-    if (closeBtn) {
-        closeBtn.click();
-        return {action: "close_button"};
-    }
-
-    // Try pressing Escape as fallback
-    document.dispatchEvent(new KeyboardEvent('keydown', {key: 'Escape', code: 'Escape'}));
-    return {action: "escape_key"};
-})();
-"""
+# Import canonical JS snippets from the game plugin (single source of truth).
+from games.breakout71.modal_handler import (  # noqa: E402
+    CLICK_PERK_JS as _CLICK_PERK_JS,
+    DETECT_STATE_JS as _DETECT_STATE_JS,
+    DISMISS_GAME_OVER_JS as _DISMISS_GAME_OVER_JS,
+    DISMISS_MENU_JS as _DISMISS_MENU_JS,
+)
 
 
 def _detect_game_state(driver) -> dict:
